assignments/week-3/Document_qa_p/requirements.txt
azure-search-documents
azure-core
azure-storage-blob
openai
python-dotenv
